requirements.txt
paho-mqtt==2.1.0
PyDect200==0.0.16
